CI/CD,Build,Packaging: autogen DimOS Ubuntu distro .img
This patch adds support for generating a .img file for DimOS.
The .img file has the DimOS source, and the dev dockers
(ros,dev,ros-python) baked into it. User should be able to simply
install the .img and spin up the dimos cli without any extra
work, at least for the non-CUDA-accelerated Dimos build.

The scripts place the dimos files in the following subdirs of the
ubuntu distro:
+ /home/ubuntu/
  + dimos-git: Dimos source, with LFS files checked out.
  + setup/: shell scripts to load docker images into docker service.
    + artifacts/: docker image tarballs.

GRUB is configured with a dimos logo for some flair and
branding.